Ignore and Kibbie folder
added in a .gitignore to ignore .DS_Store files from OS X, The Kibbie
folder is kinda unrealated but I thought it belonged here. It contains
the code needed to run the tower light bars in the kibbie dome on the
handrails, it's mostly written by Dr. Rinker and I've added in some
code for patterns and what not.